requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.41.2
aiohttp==3.9.3
